NetHandler now works for parallel connections, allowing very fast translation into multiple languages at once.
Changed storing any details intrinsic to translations to the Translation
class to help isolate them from influence of other requests. 

The Response class now returns a String result instead of an InputStream
result, offloading the stream reading to happen in it's own thread.

GoogleTranslatorAdapter fully functional for basic translation
transactions.

Multiple languages are now stored in a fast EnumSet instead of the
previous mix of Sets and Language arrays.

Developed tests for the NetHandler and GoogleTranslatorAdapter classes.

These changes broke all but the core of the plugin, but should allow for
more structured and logical development.